Integrate each Composable into an Activity or Fragment
