+++ app/src/main/res/drawable/ic_round_share_24.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M18,15.75C17.596,15.75 17.196,15.832 16.825,15.99C16.453,16.149 16.118,16.381 15.838,16.673L8.901,12.77C9.035,12.266 9.035,11.734 8.901,11.23L15.838,7.327C16.343,7.85 17.022,8.172 17.747,8.231C18.473,8.29 19.195,8.083 19.778,7.648C20.362,7.213 20.766,6.58 20.917,5.869C21.067,5.157 20.953,4.414 20.595,3.78C20.237,3.147 19.661,2.665 18.974,2.426C18.287,2.187 17.536,2.206 16.862,2.481C16.188,2.756 15.637,3.267 15.313,3.918C14.989,4.57 14.914,5.317 15.101,6.02L8.163,9.923C7.749,9.491 7.215,9.193 6.63,9.067C6.044,8.941 5.435,8.993 4.879,9.217C4.324,9.44 3.848,9.824 3.513,10.32C3.177,10.816 2.998,11.401 2.998,12C2.998,12.599 3.177,13.184 3.513,13.68C3.848,14.176 4.324,14.56 4.879,14.783C5.435,15.007 6.044,15.059 6.63,14.933C7.215,14.807 7.749,14.509 8.163,14.077L15.101,17.98C14.94,18.585 14.973,19.226 15.195,19.812C15.417,20.398 15.816,20.9 16.338,21.247C16.859,21.594 17.476,21.77 18.103,21.748C18.729,21.727 19.333,21.51 19.829,21.128C20.326,20.746 20.691,20.218 20.872,19.618C21.053,19.019 21.042,18.377 20.84,17.784C20.639,17.191 20.256,16.676 19.747,16.311C19.237,15.946 18.627,15.75 18,15.75Z"
+      android:fillColor="#E9ECEF"/>
+</vector>
