+++ app/src/main/res/drawable/ic_round_trash_24.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M20.25,4.5H15.75V3.375C15.75,2.878 15.552,2.401 15.201,2.049C14.849,1.698 14.372,1.5 13.875,1.5H10.125C9.628,1.5 9.151,1.698 8.799,2.049C8.448,2.401 8.25,2.878 8.25,3.375V4.5H3.75C3.551,4.5 3.36,4.579 3.22,4.72C3.079,4.86 3,5.051 3,5.25C3,5.449 3.079,5.64 3.22,5.78C3.36,5.921 3.551,6 3.75,6H4.547L5.438,20.293C5.504,21.552 6.469,22.5 7.688,22.5H16.313C17.537,22.5 18.483,21.573 18.563,20.297L19.453,6H20.25C20.449,6 20.64,5.921 20.78,5.78C20.921,5.64 21,5.449 21,5.25C21,5.051 20.921,4.86 20.78,4.72C20.64,4.579 20.449,4.5 20.25,4.5ZM9.027,19.5H9C8.806,19.5 8.619,19.425 8.479,19.29C8.339,19.155 8.257,18.971 8.25,18.777L7.875,8.277C7.868,8.078 7.94,7.884 8.076,7.739C8.211,7.593 8.399,7.507 8.598,7.5C8.797,7.493 8.991,7.565 9.136,7.701C9.282,7.836 9.368,8.024 9.375,8.223L9.75,18.723C9.754,18.822 9.738,18.92 9.703,19.012C9.669,19.105 9.617,19.189 9.549,19.262C9.482,19.334 9.401,19.392 9.312,19.433C9.222,19.474 9.125,19.497 9.027,19.5ZM12.75,18.75C12.75,18.949 12.671,19.14 12.53,19.28C12.39,19.421 12.199,19.5 12,19.5C11.801,19.5 11.61,19.421 11.47,19.28C11.329,19.14 11.25,18.949 11.25,18.75V8.25C11.25,8.051 11.329,7.86 11.47,7.72C11.61,7.579 11.801,7.5 12,7.5C12.199,7.5 12.39,7.579 12.53,7.72C12.671,7.86 12.75,8.051 12.75,8.25V18.75ZM14.25,4.5H9.75V3.375C9.749,3.326 9.759,3.277 9.777,3.231C9.796,3.185 9.824,3.144 9.859,3.109C9.894,3.074 9.935,3.046 9.981,3.027C10.027,3.009 10.076,2.999 10.125,3H13.875C13.924,2.999 13.973,3.009 14.019,3.027C14.065,3.046 14.106,3.074 14.141,3.109C14.176,3.144 14.204,3.185 14.223,3.231C14.241,3.277 14.251,3.326 14.25,3.375V4.5ZM15.75,18.777C15.743,18.971 15.661,19.155 15.521,19.29C15.381,19.425 15.194,19.5 15,19.5H14.973C14.874,19.497 14.778,19.474 14.688,19.433C14.598,19.392 14.518,19.333 14.45,19.261C14.383,19.189 14.331,19.105 14.297,19.012C14.262,18.92 14.246,18.822 14.25,18.723L14.625,8.223C14.628,8.125 14.651,8.028 14.692,7.938C14.733,7.849 14.791,7.768 14.864,7.701C14.936,7.634 15.02,7.581 15.113,7.547C15.205,7.512 15.303,7.496 15.402,7.5C15.5,7.504 15.597,7.526 15.687,7.567C15.776,7.608 15.857,7.666 15.924,7.739C15.991,7.811 16.044,7.895 16.078,7.988C16.113,8.08 16.128,8.178 16.125,8.277L15.75,18.777Z"
+      android:fillColor="#E9ECEF"/>
+</vector>
